Lazily load Guided Tour components only when needed.
Guided Tour components were always loaded, even when they would
not be rendered.

This commit moves the main GuidedTours logic to a new
GuidedToursComponent, which only gets loaded when guided tours
should be shown.
